stage1: diagnostic functionality for rkt run
After this patch, if the app exits with `ExecMainStatus == 203`, the app's reaper runs
the diagnostic tool and prints the output on stdout. systemd sets `ExecMainstatus`
to EXIT_EXEC (203) when execve() fails. References:
Exit status: https://github.com/systemd/systemd/blob/v230/src/basic/exit-status.h#L47
call execve(): https://github.com/systemd/systemd/blob/v230/src/core/execute.c#L1998

Example:
```
sudo build-rkt-1.9.1+git/target/bin/rkt run coreos.com/etcd:v2.0.10 --exec=/foo-not-there
image: using image from file /usr/lib/rkt/stage1-images/stage1-coreos.aci
image: using image from local store for image name coreos.com/etcd:v2.0.10
networking: loading networks from /etc/rkt/net.d
networking: loading network default with type ptp
Error: Unable to open "/foo-not-there": No such file or directory
```
Fixes https://github.com/coreos/rkt/issues/2506
