+++ stage1/diagnostic/diagnostic.c
@@ -1,0 +1,31 @@
+// Copyright 2014-2016 The rkt Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#include <unistd.h>
+
+#include "diagnostic-util.h"
+
+int main(int argc, char *argv[])
+{
+	const char *root, *exe;
+	exit_if(argc != 3, "Usage: %s /path/to/root /to/exe", argv[0]);
+
+	root = argv[1];
+	exe = argv[2];
+
+	pexit_if(chroot(root) == -1, "Chroot \"%s\" failed", root);
+	diag(exe);
+
+	return 0;
+}
