+++ stage1/diagnostic/diagnostic.mk
@@ -1,0 +1,4 @@
+ASSCB_EXTRA_HEADERS := diagnostic-util.h elf.h
+ASSCB_EXTRA_SOURCES := diagnostic-util.c
+
+include stage1/makelib/aci_simple_static_c_bin.mk
