+++ stage1/diagnostic/elf.h
