+++ stage1/makelib/aci_simple_static_c_bin.mk
@@ -2,6 +2,8 @@
 #
 # ASSCB_EXTRA_HEADERS - additional local C headers used
 #
+# ASSCB_EXTRA_SOURCES - additional local C sources used
+#
 # ASSCB_FLAVORS - which flavors want this app, set to STAGE1_FLAVORS
 # if empty
 #
@@ -37,7 +39,7 @@ endif
 # variables for build_static_c_bin.mk
 BSCB_BINARY := $(_ASSCB_BINARY_)
 BSCB_HEADERS := $(foreach h,$(ASSCB_EXTRA_HEADERS),$(MK_SRCDIR)/$h)
-BSCB_SOURCES := $(MK_SRCDIR)/$(_ASSCB_NAME_).c
+BSCB_SOURCES := $(MK_SRCDIR)/$(_ASSCB_NAME_).c $(foreach h,$(ASSCB_EXTRA_SOURCES),$(MK_SRCDIR)/$h)
 BSCB_ADDITIONAL_CFLAGS := -Wall -Os $(ASSCB_EXTRA_CFLAGS)
 
 CLEAN_FILES += $(_ASSCB_BINARY_)
