+++ stage1/secondary-stuff.mk
@@ -3,6 +3,7 @@ _S1_SS_SUBDIRS_ := \
 	enter \
 	enter_kvm \
 	enterexec \
+	diagnostic \
 	gc \
 	init \
 	net \
